requirements.txt
asyncpg==0.25.0
aioredis==2.0.1
aiopg==1.3.3
SQLAlchemy==1.4.36
databases==0.5.5
fastapi==0.78.0
python-dotenv==0.20.0
pydantic==1.9.1
uvicorn==0.17.6